pyproject.toml
[tool.poetry]
name = "git_project_status"
version = "0.0.13"
description = "Check subdirectories for dirty repos."
authors = ["James Hodgkinson <james@terminaloutcomes.com>"]
license = "MIT"
readme = "README.md"
keywords = [
    "git"
]
classifiers = [
        'Development Status :: 5 - Production/Stable',
        'Intended Audience :: Developers',
        'Topic :: Internet :: WWW/HTTP :: Dynamic Content',
        'License :: OSI Approved :: MIT License',
        'Programming Language :: Python :: 3',
        'Programming Language :: Python :: 3.8',
        'Programming Language :: Python :: 3.9',
        'Programming Language :: Python :: 3.10',
]

[tool.poetry.dependencies]
python = "^3.8"
loguru = "^0.6.0"
GitPython = "^3.1.27"
click = "^8.1.3"

[tool.poetry.dev-dependencies]
pylint = "^2.15.3"
pytest = "^7.1.3"
mypy = "^0.982"
black = "^22.10.0"


[tool.poetry.scripts]
git_project_status = "git_project_status.__main__:cli"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.MASTER]
max-line-length=200
disable="W0511"
